Add Delete Functionality

+++ src/user/user.controller.ts
@@ -1,9 +1,10 @@
-import { Controller, Get, Param, Delete, UseGuards, Req, ConflictException } from '@nestjs/common';
+import { Controller, Get, Param, Delete, UseGuards, Req, ConflictException, NotFoundException } from '@nestjs/common';
 import { UserService } from './user.service';
 import { AuthGuard } from 'src/auth/guard/auth.guard';
 import { RequiredRole } from 'src/auth/decorator/required-role.decorator';
 import { Role } from './roles.enum';
 import { RoleGuard } from 'src/auth/guard/role.guard';
+import { isValidObjectId } from 'mongoose';
 
 @UseGuards(AuthGuard, RoleGuard)
 @RequiredRole(Role.administrator)
@@ -30,8 +31,15 @@ export class UserController {
     if (request.user.id === id) {
       throw new ConflictException('You are not allowed to delete yourself')
     }
+    
+    if (!isValidObjectId(id)) {
+      throw new ConflictException('Invalid User-Id');
+    }
 
-    // TODO: Delete all Decks and Cards
+    const user = await this.userService.findOne(id);
+    if (!user) {
+      throw new NotFoundException(`User with Id ${ id } not found`)
+    }
 
     return this.userService.remove(id);
   }
